Refactor quiz logic and UI in ViewController.swift
Added outlets for every UI element. Create variables to store global data. Create an array of questions using a 2D Array. Set a new question when the view loads. Handle quiz logic when either button is pressed to display questions and keep track of the score. Use the progress bar and set the progress using the outlet and calculating the percerntage of the progress